requirements.txt
altgraph==0.17
future==0.18.2
pefile==2019.4.18
Pillow==8.0.1
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.10
pywin32-ctypes==0.2.0
